+++ cmd/root.go
@@ -35,6 +35,7 @@ func init() {
 	rootCmd.Flags().Lookup("version").Shorthand = "v"
 	rootCmd.AddCommand(runCmd)
 	rootCmd.AddCommand(initCmd)
+	rootCmd.AddCommand(scaffoldCmd)
 	rootCmd.AddCommand(switchCmd)
 	rootCmd.AddCommand(revertCmd)
 }